Build fix.
svn:sync-xref-src-repo-rev: r126137

+++ src/VBox/Main/src-client/DrvAudioVideoRec.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvAudioVideoRec.cpp 75082 2018-10-25 16:52:01Z andreas.loeffler@oracle.com $ */
+/* $Id: DrvAudioVideoRec.cpp 75083 2018-10-25 16:59:14Z andreas.loeffler@oracle.com $ */
 /** @file
  * Video recording audio backend for Main.
  */
@@ -769,7 +769,7 @@ static DECLCALLBACK(int) drvAudioVideoRecStreamPlay(PPDMIHOSTAUDIO pInterface, P
             {
                 memcpy((uint8_t *)pStreamAV->pvSrcBuf + cbSrc, pvCircBuf, cbCircBuf);
 
-                cbSrc += cbCircBuf;
+                cbSrc += (uint32_t)cbCircBuf;
                 Assert(cbSrc <= pStreamAV->cbSrcBuf);
             }
 
@@ -811,7 +811,7 @@ static DECLCALLBACK(int) drvAudioVideoRecStreamPlay(PPDMIHOSTAUDIO pInterface, P
             pSink->Codec.STAM.msEncTotal += pSink->Codec.Opus.msFrame * cEncFrames;
 # endif
             Assert((uint32_t)cbWritten <= (uint32_t)pStreamAV->cbDstBuf);
-            const uint32_t cbDst = RT_MIN((uint32_t)cbWritten, pStreamAV->cbDstBuf);
+            const uint32_t cbDst = RT_MIN((uint32_t)cbWritten, (uint32_t)pStreamAV->cbDstBuf);
 
             Assert(cEncFrames == 1);
